Separatinf input system from core. Making opengl render system implementation into Pimpl to hide private members dependent on SDL/Glew headers
