+++ include/systems/IRenderSystem.h
@@ -28,15 +28,15 @@ class IRenderSystem: public UListener<core_events::IPreRenderListener, core_even
 public:
     virtual ~IRenderSystem() = default;
 
-    [[nodiscard]] virtual bool Initialize(std::uint32_t width, std::uint32_t heigth) = 0;
+    [[nodiscard]] virtual bool Initialize() = 0;
     [[nodiscard]] virtual std::unique_ptr<SShaderHandle> CompileShader(const char *vertexShaderSrc, const char *fragmentShaderSrc) const = 0;
-    [[nodiscard]] virtual std::unique_ptr<SRenderBufferHandle> WriteRenderBufferData(const unboxing_engine::CMeshBuffer &meshBuffer) const = 0;
-    [[nodiscard]] virtual void EraseRenderBufferData(const SRenderBufferHandle &renderBufferHandle) const = 0;
+    [[nodiscard]] virtual std::unique_ptr<SRenderBufferHandle> WriteRenderBufferData(const unboxing_engine::CMeshBuffer &meshBuffer) = 0;
+    [[nodiscard]] virtual void EraseRenderBufferData(const SRenderBufferHandle &renderBufferHandle) = 0;
     [[nodiscard]] virtual const Camera &GetCamera() const = 0;
     [[nodiscard]] virtual const SShaderHandle &GetShader(std::uint32_t id) const = 0;
-    virtual void GetCamera(const Camera& camera) = 0;
+    virtual void SetCamera(const Camera& camera) = 0;
 
-    virtual void Render(const SRenderContextHandle &renderContextHandle) const = 0;
+    virtual void Render(const SRenderContextHandle &renderContextHandle) = 0;
 };
 
 }// namespace unboxing_engine
