+++ sdl_gl_render_system/sdl_input_system_lib.h
@@ -1,0 +1,24 @@
+#pragma once
+
+#include <systems/IInputSystem.h>
+#include <CoreEvents.h> 
+
+namespace unboxing_engine::systems {
+class CSDLInputSystem : public IInputSystem {
+public:
+    ~CSDLInputSystem() override;
+
+    void OnInput() override;
+
+private:
+    bool OnMouseInput();
+    bool OnKeyboardInput();
+
+    core_events::SCursor mCursor;
+    core_events::SKeyboard mKeyboard;
+
+    ///Keyboard input attributes
+    std::uint8_t const *keyState = nullptr;
+    bool quit = false;
+};
+}// namespace unboxing_engine::systems
